Unify diff parsing/rendering utilities (#611) (#680)
* Unify diff parsing/rendering utilities (#611)

Extract shared diff parsing and rendering logic into src/lib/diff/ module.

- Created parseDetailedDiff() for workspace diff viewer with line numbers
- Created parseFileDiff() for PR detail panel with file-based hunks
- Shared diff styling utilities for consistent visual output
- Removed ~160 lines of duplicated code
- All tests passing (1300/1300)


* Fix diff parsing correctness issues from Greptile review

Address two edge cases in diff parsing:

1. parseDetailedDiff: Only assign line numbers when inside a hunk
   - Prevents invalid line number 0 if additions/deletions appear before
     the first @@ hunk header
   - Tracks hunk state with `inHunk` flag

2. parseFileDiff: Only treat space-prefixed lines as context
   - Prevents corruption of metadata lines (e.g., "\ No newline at end
     of file") that don't start with space
   - Explicitly handles empty lines and space-prefixed lines separately
   - Skips non-context metadata lines instead of mangling them

All tests passing (1300/1300).


---------
